Consolidate Drive interfacing tiles into module
The logic for handling multiple drives and their display slots is now held inside of DriveInterfaceModule which also leads to plenty of performance improvements as we do not need to query the drives every tick but rather only when the inventory updates which is much less frequently, noticeable when there are many tiles loaded at once.

TODO: Make recipe selection buttons and the crafting of the fabricator work.